Build a Python module
