+++ include/lemon/launch.hpp
@@ -7,6 +7,13 @@
 
 namespace lemon {
 
+class LemonPythonBase : public boost::noncopyable {
+public:
+    virtual ~LemonPythonBase() {}
+    virtual std::string worker(chemfiles::Frame&, const std::string&) = 0;
+    virtual void finalize() = 0;
+};
+
 template<typename Map1, typename Map2>
 struct map_combine {
     void operator()(Map1& map1, const Map2& map2) const {
